# Optimize JSON escaping