Instead of using multiple gsub! calls, this uses binary encoding to do the regexp gsub with map, then forces encoding back to UTF-8.

This uses the same approach as before ebe0c4066b5b2a4b0e08220b04b33243b54d8e34, with binary encoding, which performs better.


```ruby
# frozen_string_literal: true

require "bundler/inline"

gemfile do
  source 'https://rubygems.org'
  gem 'benchmark-ips'
  gem 'rails', path: '.'
  gem "json", path: "../../ruby/json"
  gem "vernier"
end

require "active_support"
require "active_support/json/encoding"

obj = JSON.load_file("../../ruby/json/benchmark/data/twitter.json")

Benchmark.ips do |x|
  x.report "obj.to_json" do
    obj.to_json
  end
end

obj = "a"*obj.to_json.bytesize
Benchmark.ips do |x|
  x.report "obj.to_json nothing to escape" do
    obj.to_json
  end
end
```

Before:
```
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
         obj.to_json    28.000 i/100ms
Calculating -------------------------------------
         obj.to_json    290.133 (± 2.4%) i/s    (3.45 ms/i) -      1.456k in   5.020937s
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
obj.to_json nothing to escape
                       100.000 i/100ms
Calculating -------------------------------------
obj.to_json nothing to escape
                          1.027k (± 3.7%) i/s  (973.57 μs/i) -      5.200k in   5.069529s
```

After:
```
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
         obj.to_json    61.000 i/100ms
Calculating -------------------------------------
         obj.to_json    610.559 (± 4.6%) i/s    (1.64 ms/i) -      3.050k in   5.007098s
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
obj.to_json nothing to escape
                        97.000 i/100ms
Calculating -------------------------------------
obj.to_json nothing to escape
                          1.004k (± 2.4%) i/s  (996.13 μs/i) -      5.044k in   5.027348s
```

We can see it's a bit slower without anything to escape, but within the margin of error. We also think we can optimize gsub with a map on the Ruby side to avoid the allocations mentioned in https://github.com/rails/rails/pull/48669. cc @jhawthorn