@byroot 2025-02-11
> We also think we can optimize gsub with a map on the Ruby side to avoid the allocations mentioned in #48669.

Ruby PR to get rid of the string allocations: https://github.com/ruby/ruby/pull/12730, but I'll try to follow it up to try to also reduce `MatchData` allocations.